requirements.txt
altgraph==0.17.4
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cx_Logging==3.2.1
cycler==0.12.1
filelock==3.17.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
lap==0.5.12
lief==0.16.3
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.1
onnx==1.17.0
onnxruntime-gpu==1.20.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMySQL==1.1.1
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==2.3.0
